security-package/issues/62: ReCaptcha refactoring according to Magento Technical Guidelines
